+++ android/src/main/java/com/reactjjkit/cropper/CropperModule.kt
@@ -2,9 +2,12 @@ package com.reactjjkit.cropper
 
 import android.graphics.*
 import android.util.Base64
+import androidx.core.content.ContextCompat
+import androidx.core.graphics.drawable.toBitmap
 import androidx.core.graphics.toRect
 import com.facebook.react.bridge.*
 import java.io.ByteArrayOutputStream
+import java.net.URL
 import kotlin.math.min
 import kotlin.math.round
 
@@ -23,7 +26,7 @@ class CropperModule(context: ReactApplicationContext) : ReactContextBaseJavaModu
 
 
     @ReactMethod
-    fun makeCrop(map:ReadableMap?,promise: Promise){
+    fun makeCrop64(map:ReadableMap?,promise: Promise){
         if(map != null){
             val image = map.getString("image")
             val imgRect = map.getMap("rect")
@@ -78,6 +81,67 @@ class CropperModule(context: ReactApplicationContext) : ReactContextBaseJavaModu
         }
     }
 
+
+
+    @ReactMethod
+    fun makeCropStatic(map:ReadableMap?,promise: Promise){
+        if(map != null){
+            val image = map.getString("image")
+            val imgRect = map.getMap("rect")
+            val cw = map.getDouble("cw").toFloat()
+            val ch = map.getDouble("ch").toFloat()
+            val crop = map.getMap("crop")
+            val wr = map.getInt("width")
+            val hr = map.getInt("height")
+            val quality = map.getDouble("quality").toFloat()
+            val format = map.getInt("format")
+            val bmp : Bitmap?
+
+            bmp = if(image != null && image.contains("http")) {
+                val url = URL(image)
+                BitmapFactory.decodeStream(url.openConnection().getInputStream())
+            }else{
+                val id = reactContext.resources.getIdentifier(image,"drawable",reactContext.packageName)
+                ContextCompat.getDrawable(reactContext,id)?.toBitmap()
+            }
+            if(image != null && imgRect != null && crop != null && bmp != null){
+
+                val r = mapToRectF(imgRect)
+                val c = mapToRectF(crop)
+
+                val rf = CropperHelper.crop(bmp,r,cw,ch,c)
+                val bf = cropBitmap(bmp,rf)
+                val fmt = if(format == 0) Bitmap.CompressFormat.JPEG else Bitmap.CompressFormat.PNG
+                if(wr > 0 && hr > 0){
+                    val br = fitCenter(bf,wr,hr)
+                    val output = ByteArrayOutputStream()
+                    if( br.compress(fmt,(quality*100).toInt(),output) ){
+                        val array = output.toByteArray()
+                        val encodedString = Base64.encodeToString(array,Base64.DEFAULT)
+                        promise.resolve(encodedString)
+                    }else{
+                        promise.resolve(null)
+                    }
+
+                }else{
+                    val output = ByteArrayOutputStream()
+                    if( bf.compress(fmt,(quality*100).toInt(),output) ){
+                        val array = output.toByteArray()
+                        val encodedString = Base64.encodeToString(array,Base64.DEFAULT)
+                        promise.resolve(encodedString)
+                    }else{
+                        promise.resolve(null)
+                    }
+                }
+
+            }else{
+                promise.resolve(null)
+            }
+        }else{
+            promise.resolve(null)
+
+        }
+    }
 
     private fun cropBitmap(bmp:Bitmap,rect:RectF) : Bitmap {
         val r = rect.toRect()