+++ utils/Rect.js
@@ -1,7 +1,7 @@
 
 const centerRect = (width,height,cw,ch) => {
     if(width == cw && height == ch){
-        return { left: 0, top: 0, right:left + cw, bottom: top + ch }
+        return { left: 0, top: 0, right: cw, bottom: ch }
     }
     
     let l = cw / 2 - width/2
@@ -11,22 +11,35 @@ const centerRect = (width,height,cw,ch) => {
 }
 
 const fitCenterRect = (width,height,cw,ch)=>{
-    if (width > height) {
-        let h = height / width  * cw
-        let dy = ch / 2 - h/2
-        return {left: 0, top: dy, right: cw, bottom: dy + h}
-    }else{
-        let w = width / height  * ch
-        let dx = cw / 2 - w/2
-        return {left: dx, top: 0, right:dx + w, bottom: ch}
+    if (cw == width && ch == height) {
+        return {left: 0, top: 0, right:width, bottom: height}
     }
+    var  widthPercentage = cw /  width
+    var  heightPercentage = ch /  height
+    var  minPercentage = Math.min(widthPercentage, heightPercentage)
+
+    var targetWidth = Math.round(minPercentage * width)
+    var targetHeight = Math.round(minPercentage * height)
+
+    if (width == targetWidth && height == targetHeight) {
+         return {left: 0, top: 0, right:width, bottom: height}
+    }
+    targetWidth = (minPercentage * width) | 0
+    targetHeight = (minPercentage * height) | 0
+   
+    var dx = (cw - targetWidth) / 2
+    var dy = (ch - targetHeight) / 2
+
+     return {left: dx, top: dy, right:dx + targetWidth, bottom: dy + targetHeight}
+    
 }
 
 const offset = (rect,dx,dy)=>{
-    rect.left += dx
-    rect.top += dy
-    rect.right += dx
-    rect.bottom += dy
+    let l = rect.left + dx
+    let t = rect.top + dy
+    let r = rect.right + dx
+    let b = rect.bottom + dy
+    return { left: l, top: t, right: r, bottom: b}
 }
 
 // pivot center
@@ -38,10 +51,11 @@ const scale = (rect,scale) => {
         let nh = h * scale
         let dx = (w - nw) / 2
         let dy = (h - nh) / 2
-        rect.left += dx
-        rect.top += dy
-        rect.right -= dx
-        rect.bottom -= dy
+        let l = rect.left + dx
+        let t = rect.top + dy
+        let r = rect.right - dx
+        let b = rect.bottom - dy
+        return { left: l, top: t, right: r, bottom: b}
     }
 }
 
@@ -50,50 +64,73 @@ const scale = (rect,scale) => {
 // }
 
 const inset = (rect,dx,dy)=>{
-    rect.left += dx
-    rect.top += dy
-    rect.right -= dx
-    rect.bottom -= dy
+    let l = rect.left + dx
+    let t = rect.top + dy
+    let r = rect.right - dx
+    let b = rect.bottom - dy
+    return { left: l, top: t, right: r, bottom: b}
 }
 
 const insetBr = (rect,distance) => {
-    rect.right -= distance
-    rect.bottom -= distance
+    let l = rect.left 
+    let t = rect.top 
+    let r = rect.right - distance
+    let b = rect.bottom - distance 
+    return { left: l, top: t, right: r, bottom: b}
 }
 const insetBl = (rect,distance) => {
-    rect.left += distance
-    rect.bottom -= distance
+    let l = rect.left + distance
+    let t = rect.top 
+    let r = rect.right 
+    let b = rect.bottom - distance
+    return { left: l, top: t, right: r, bottom: b}
 }
 
 const insetTl = (rect,distance) => {
-    rect.left += distance
-    rect.top += distance
+    let l = rect.left + distance
+    let t = rect.top + distance
+    let r = rect.right 
+    let b = rect.bottom 
+    return { left: l, top: t, right: r, bottom: b}
 }
 const insetTr = (rect,distance) => {
-    rect.right -= distance
-    rect.top += distance
+    let l = rect.left 
+    let t = rect.top + distance
+    let r = rect.right - distance
+    let b = rect.bottom 
+    return { left: l, top: t, right: r, bottom: b}
 }
 const insetBlr = (rect,distance) => {
-    rect.left += distance
-    rect.bottom -= distance
-    rect.right -= distance
+    let l = rect.left + distance
+    let t = rect.top 
+    let r = rect.right - distance
+    let b = rect.bottom - distance
+    return { left: l, top: t, right: r, bottom: b}
 }
 const insetTlr = (rect,distance) => {
-    rect.left += distance
-    rect.top += distance
-    rect.right -= distance
+    let l = rect.left + distance
+    let t = rect.top + distance
+    let r = rect.right - distance
+    let b = rect.bottom 
+    return { left: l, top: t, right: r, bottom: b}
 }
 
 const insetLtb = (rect,distance) => {
-    rect.left += distance
-    rect.top += distance
-    rect.bottom -= distance
+    let l = rect.left + distance
+    let t = rect.top + distance
+    let r = rect.right
+    let b = rect.bottom - distance
+    return { left: l, top: t, right: r, bottom: b}
+
 }
 
 const insetRtb = (rect,distance) => {
-    rect.right -= distance
-    rect.top += distance
-    rect.bottom -= distance
+    let l = rect.left
+    let t = rect.top + distance
+    let r = rect.right - distance
+    let b = rect.bottom - distance
+    return { left: l, top: t, right: r, bottom: b}
+ 
 }