+++ readmes/changelog.md
@@ -1,5 +1,13 @@
 # Change Log
 
+**1.6.0**
+
+- [Cropper](./readmes/cropper.md) 
+    - Renamed makeCrop to makeCrop64
+    - added makeCropStatic for static images.
+-  Util Rect, now every function return a rect.
+
+
 **1.5.9**
 
 - added  [Cropper](./readmes/cropper.md)